move table sample to planner util
