+++ pkg/planner/core/BUILD.bazel
@@ -122,6 +122,7 @@ go_library(
         "//pkg/planner/util/debugtrace",
         "//pkg/planner/util/fixcontrol",
         "//pkg/planner/util/optimizetrace",
+        "//pkg/planner/util/tablesample",
         "//pkg/privilege",
         "//pkg/sessionctx",
         "//pkg/sessionctx/stmtctx",
