+++ pkg/planner/util/tablesample/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "tablesample",
+    srcs = ["sample.go"],
+    importpath = "github.com/pingcap/tidb/pkg/planner/util/tablesample",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/expression",
+        "//pkg/parser/ast",
+        "//pkg/table",
+        "//pkg/util/size",
+    ],
+)
